% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anthesis.R
\name{anthesis}
\alias{anthesis}
\alias{anthesis.default}
\alias{anthesis.array}
\alias{anthesis.clima_ls}
\alias{anthesis.sf}
\title{Crop sensitive indices during anthesis}
\source{
Challinor et al. (2016). Nature Climate Change 6(10):6954-958
\cr\url{https://doi.org/10.1038/nclimate3061}

Trnka et al. (2014). Nature Climate Change 4(7):637–43.
\cr\url{https://doi.org/10.1038/nclimate2242}
}
\usage{
anthesis(object, ...)

\method{anthesis}{default}(object, day.one, ...)

\method{anthesis}{array}(object, day.one, ...)

\method{anthesis}{clima_ls}(object, ...)

\method{anthesis}{sf}(object, day.one, ..., as.sf = TRUE)
}
\arguments{
\item{object}{a data.frame (or any object that can be coerced to data.frame) 
with geographical coordinates (lonlat), or an object of class \code{sf}
with geometry 'POINT' or 'POLYGON', or an \code{array} with two dimensions 
containing the temperature data or a \code{clima_ls} with maximum and minimum 
temperature, in that order. See details}

\item{...}{additional arguments passed to methods. See details}

\item{day.one}{a vector of class \code{Date} or any other object that can be 
coerced to \code{Date} (e.g. integer, character YYYY-MM-DD) for the starting 
day to capture the climate data}

\item{as.sf}{logical, to return an object of class 'sf'}
}
\value{
A dataframe with crop sensitive indices with n colunms depending on the 
 number of thresholds passed to each index:
\item{hts_mean}{high temperature stress using daily MEAN temperature, 
 and given as percentage number of days a certain threshold is exceeded}
\item{hts_max}{high temperature stress using daily MAX temperature,
 and given as percentage number of days a certain threshold is exceeded}
\item{hse}{heat stress event, and given as percentage number of days a 
 a certain threshold is exceeded for at least two consecutive days}
\item{hse_ms}{heat stress event, and given the maximum number of days 
 a certain threshold is exceeded for at least two consecutive days}
}
\description{
Compute crop sensitive indices. These indices are intended to capture 
 the changes in temperature extremes during anthesis, but can also be 
 applied to other phenological stages.
}
\details{
The function uses pre-defined threshold to compute the indices. For hts_mean (32,
 35, 38 Celsius), for hts_max (36, 39, 42 Celsius) and for hse (31 Celsius). 
 See Additional arguments 

Additional arguments:

The thresholds can be adjusted using the arguments \code{hts_mean.threshold}, 
 \code{hts_max.threshold} and \code{hse.threshold}, which are a numeric (or
 vector of numeric)

\code{last.day}: an object (optional to \var{span}) of class \code{Date} or
 any other object that can be coerced to \code{Date} (e.g. integer, character 
 YYYY-MM-DD) for the last day of the time series
 
\code{span}: an integer (optional to \var{last.day}) or a vector with 
 integers (optional if \var{last.day} is given) for the length of 
 the time series to be captured
}
\examples{
data("modis", package = "climatrends")
 
anthesis(modis, 
         day.one = "2013-10-27", 
         last.day = "2013-11-04", 
         hse.threshold = c(33.5))

}
\seealso{
Other temperature functions: 
\code{\link{ETo}()},
\code{\link{GDD}()},
\code{\link{temperature}()}
}
\concept{temperature functions}
