% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/climatrends.R
\docType{package}
\name{climatrends}
\alias{climatrends}
\alias{climatrends-package}
\title{Climate Variability Indices for Ecological Modelling}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

Supports analysis of trends in climate change, ecological and crop modelling.
}
\seealso{
\strong{Useful links:}
\itemize{
\item{Development repository: 
 \url{https://github.com/agrobioinfoservices/climatrends}}
\item{Static documentation: 
 \url{https://agrobioinfoservices.github.io/climatrends/}}
\item{Report bugs: 
 \url{https://github.com/agrobioinfoservices/climatrends/issues}}
}
}
\author{
Kauê de Sousa and Jacob van Etten and Svein Ø. Solberg
}
