% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crop_sensitive.R
\name{crop_sensitive}
\alias{crop_sensitive}
\title{Crop sensitive indices}
\source{
Challinor et al. (2016). Nature Climate Change, 6, 954-958(2016)
\cr\url{https://doi.org/10.1038/nclimate3061}
}
\usage{
crop_sensitive(object, ...)
}
\arguments{
\item{object}{a data.frame (or any object that can be coerced to data.frame) 
with geographical coordinates (lonlat), or an object of class \code{sf}
with geometry 'POINT' or 'POLYGON', or an \code{array} with two dimensions 
containing the temperature data or a \code{clima_ls} with day and night 
temperature, in that order. See details}

\item{...}{additional arguments passed to methods. See details}
}
\description{
Compute crop sensitive indices
}
