% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ETo.R
\name{ETo}
\alias{ETo}
\alias{ETo.default}
\alias{ETo.array}
\alias{ETo.sf}
\title{Reference evapotranspiration}
\usage{
ETo(object, day.one, Kc = 1, ...)

\method{ETo}{default}(object, day.one, Kc = 1, ...)

\method{ETo}{array}(object, day.one, Kc = 1, lat = NULL, p = 0.27, ...)

\method{ETo}{sf}(object, day.one, Kc = 1, as.sf = TRUE, ...)
}
\arguments{
\item{object}{a data.frame (or any object that can be coerced to data.frame) 
with geographical coordinates (lonlat), or an object of class \code{sf}
with geometry 'POINT' or 'POLYGON', or an \code{array} with two dimensions 
containing the temperature data or a \code{clima_ls} with maximum and minimum 
temperature, in that order. See details}

\item{day.one}{a vector of class \code{Date} or any other object that can be 
coerced to \code{Date} (e.g. integer, character YYYY-MM-DD) for the starting 
day to capture the climate data}

\item{Kc}{a numeric value for the crop factor for water requirement}

\item{...}{additional arguments passed to methods. See details}

\item{lat}{a vector for the latitude (in Decimal degrees), used to compute 
mean daily percentage of annual daytime hours based on the latitude and month.
This is extracted automatically in the \code{sf} method. See details}

\item{p}{optional if \var{lat} is given, a numeric for the mean daily percentage 
of annual daytime hours (p = 0.27 by default)}

\item{as.sf}{logical, to return an object of class 'sf'}
}
\value{
The evapotranspiration in mm/day
}
\description{
Reference evapotranspiration using the Blaney-Criddle method. This is general  
theoretical method used when no measured data on pan evaporation 
is available locally.
}
\details{
When \var{lat} is provided, it is combined with the month provided in 
 \var{day.one} to call for the system data \code{daylight} to find
 the correct value for \var{p} which represents the daily percentage
 of daytime hours in the given month and latitude. Otherwise \var{p} is set 
 to 0.27 as default.
 
The \code{array} method assumes that \var{object} contains climate data provided 
 from a local source; this requires an array with two dimensions, 1st dimension 
 contains the day temperature and 2nd dimension the night temperature, 
 see help("modis", "climatrends") for an example on input structure.

The \code{default} method and the \code{sf} method assumes that the climate data
 will e fetched from a remote (cloud) source that be adjusted using the argument 
 \var{data.from}.

Additional arguments:

\code{last.day}: an object (optional to \var{span}) of class \code{Date} or
 any other object that can be coerced to \code{Date} (e.g. integer, character 
 YYYY-MM-DD) for the last day of the time series
 
\code{span}: an integer (optional to \var{last.day}) or a vector with 
 integers (optional if \var{last.day} is given) for the length of 
 the time series to be captured

\code{data.from}: character for the source of climate data. Current remote data 
 is: 'nasapower'

\code{pars}: character vector for the temperature data to be fetched. If 
 \code{data.from} is 'nasapower'. The temperature can be adjusted to 2 m, the default,
 c("T2M_MAX", "T2M_MIN") or 10 m c("T10M_MAX", "T10M_MIN") 

\code{days.before}: optional, an integer for the number of days before 
 \var{day.one} to be included in the timespan.
}
\examples{
# Using local data
data("modis", package = "climatrends")

ETo(modis, 
    day.one = "2013-10-28",
    span = 10,
    Kc = 0.92)

\donttest{
######################################

# Using remote data
# random geographic locations around bbox(11, 12, 55, 58)
set.seed(826128)
lonlat <- data.frame(lon = runif(2, 11, 12),
                     lat = runif(2, 55, 58))

# the evapotranspiration in the 30 days
ETo(lonlat,
    day.one = "2018-05-15",
    last.day = "2018-06-15",
    Kc = 0.5)

#######################################

# Objects of class 'sf'
data("lonlatsf", package = "climatrends")

ETo(lonlatsf,
    day.one = "2015-04-22",
    last.day = "2015-06-22",
    pars = c("T10M_MAX", "T10M_MIN"))
}
}
\references{
Brouwer C. & Heibloem M. (1986). Irrigation water management: 
Irrigation water needs. Food and Agriculture Organization of The 
United Nations, Rome, Italy. 
\url{http://www.fao.org/3/S2022E/s2022e00.htm}
}
\seealso{
Other temperature functions: 
\code{\link{GDD}()},
\code{\link{anthesis}()},
\code{\link{temperature}()}
}
\concept{temperature functions}
