% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_timeseries.R
\name{get_timeseries}
\alias{get_timeseries}
\alias{get_timeseries.default}
\alias{get_timeseries.sf}
\alias{get_timeseries.matrix}
\title{Time series climate data}
\usage{
get_timeseries(object, day.one = NULL, span = 150, ...)

\method{get_timeseries}{default}(object, day.one = NULL, span = 150, source = "nasapower", ...)

\method{get_timeseries}{sf}(
  object,
  day.one = NULL,
  span = 150,
  source = "nasapower",
  as.sf = FALSE,
  ...
)

\method{get_timeseries}{matrix}(object, day.one = NULL, span = 150, ...)
}
\arguments{
\item{object}{a data.frame (or object that can be coerced to data.frame) with
geographical coordinates (lonlat), or an object of class \code{sf} with 
geographical coordinates (lonlat), or a named \code{matrix} with climate 
data. See details.}

\item{day.one}{a vector of class \code{Date} for the starting date to capture 
the climate data}

\item{span}{an integer or a vector with integers for the duration of the 
time series to be captured}

\item{...}{additional arguments passed to methods. See details.}

\item{source}{character, for the source of climate data. See details.}

\item{as.sf}{logical, returns an object of class \code{[sf]{sf}}}
}
\value{
An object with time series climate data for the chosen period
}
\description{
Concatenate time series climate data
}
\details{
The default method and the sf method assumes that the climate data will be retrieved 
 from an external \code{source}.

The matrix method assumes that the climate data was previously handled and will be 
 inputted in the format of a named matrix. See help(modis) for examples.

Current remote \code{source} is: 'nasapower'

Additional arguments:

\code{pars}: character vector of solar, meteorological or climatology parameters 
to download. See help("parameters", "nasapower").

\code{days.before}: optional, an integer for the number of days before 
 \code{day.one} to be included in the timespan.
}
\examples{
# Using local sources
data("modis", package = "climatrends")

day <- as.Date("2013-10-28", format = "\%Y-\%m-\%d")

get_timeseries(chirp, day, span = c(10,11))


########################################################
\donttest{
# Fetch data from NASA POWER using 'sf' method
library("sf")
set.seed(123)
lonlat <- data.frame(lon = runif(2, 11, 12),
                     lat = runif(2, 55, 58))

lonlat <- st_as_sf(lonlat, coords = c("lon","lat"), crs = 4326)

do <- as.Date(17667, origin = "1970-01-01")

g <- get_timeseries(lonlat, day.one = do, span = 10, pars = "PRECTOT")

g

g <- get_timeseries(lonlat, day.one = do, span = c(10, 13), 
                    pars = "T10M", as.sf = TRUE)

g
}
}
\concept{GET functions}
