% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rainfall.R
\name{rainfall}
\alias{rainfall}
\alias{rainfall.default}
\alias{rainfall.matrix}
\alias{rainfall.sf}
\alias{rainfall.clima_df}
\title{Rainfall indices}
\usage{
rainfall(object, ...)

\method{rainfall}{default}(object, day.one, span = NULL, timeseries = FALSE, intervals = 5, ...)

\method{rainfall}{matrix}(object, day.one, span = NULL, timeseries = FALSE, intervals = 5, ...)

\method{rainfall}{sf}(
  object,
  day.one,
  span = NULL,
  timeseries = FALSE,
  intervals = 5,
  as.sf = TRUE,
  ...
)

\method{rainfall}{clima_df}(object, timeseries = FALSE, intervals = 5, ...)
}
\arguments{
\item{object}{a \code{data.frame} (or any other object that can be coerced to 
data.frame) with geographical coordinates (lonlat), or an object of class 
\code{sf} with geometry 'POINT' or 'POLYGON', or a named \code{matrix} with 
precipitation data, or a data.frame of class \code{clima_df}. See details.}

\item{...}{additional arguments passed to methods. See details.}

\item{day.one}{a vector of class \code{Date} or any other object that can be 
coerced to \code{Date} (e.g. integer, character YYYY-MM-DD) for the starting 
day to capture the climate data}

\item{span}{an integer or a vector with integers (optional if \var{last.day} is 
given) for the length of the time series to be captured}

\item{timeseries}{logical, \code{FALSE} for a single point time series
observation or \code{TRUE} for a time series based on \var{intervals}}

\item{intervals}{integer (no lower than 5), for the days intervals when
\var{timeseries} = \code{TRUE}}

\item{as.sf}{logical, to return an object of class 'sf'}
}
\value{
A dataframe with rainfall indices:
\item{MLDS}{maximum length of consecutive dry day, rain < 1 mm (days)}
\item{MLWS}{maximum length of consecutive wet days, rain >= 1 mm (days)}
\item{R10mm}{number of heavy precipitation days 10 >= rain < 20 mm (days)}
\item{R20mm}{number of very heavy precipitation days rain >= 20 (days)}
\item{Rx1day}{maximum 1-day precipitation (mm)}
\item{Rx5day}{maximum 5-day precipitation (mm)}
\item{R95p}{total precipitation when rain > 95th percentile (mm)}
\item{R99p}{total precipitation when rain > 99th percentile (mm)}
\item{Rtotal}{total precipitation (mm) in wet days, rain >= 1 (mm)}
\item{SDII}{simple daily intensity index, total precipitation divided by the
 number of wet days (mm/days)}
}
\description{
Methods to compute rainfall indices over a time series
}
\details{
The \code{matrix} method assumes that \var{object} contains climate data provided 
 from a local source; see help("chirp", package = "climatrends") for an example on input 
 structure.

The \code{default} method and the \code{sf} method assumes that the climate data
 will e fetched from a remote (cloud) source that be adjusted using the argument 
 \var{data.from}.

When \var{timeseries} = \code{TRUE}, an id is created, 
 which is the index for the rownames of the inputted \var{object}.

Additional arguments:

\code{last.day}: optional to \var{span}, an object of class \code{Date} or
 any other object that can be coerced to \code{Date} (e.g. integer, character 
 YYYY-MM-DD)  for the last day of the time series

\code{data.from}: character for the source of remote data. Current remote source 
 is: 'nasapower'

\code{pars}: character vector for the precipitation data to be fetched. If 
 \code{data.from} is 'nasapower', the default precipitation parameter is "PRECTOT".

\code{days.before}: optional, an integer for the number of days before 
 \var{day.one} to be included in the timespan.
 
When \var{timeseries} = \code{TRUE}, an id is created, 
 which is the index for the rownames of the inputted \var{object}.
}
\examples{
# Using local sources
data("chirp", package = "climatrends")

rainfall(chirp,
         day.one = "2013-10-28",
         span = 12)

#####################################################
\donttest{
# Using remote sources of climate data

data("lonlatsf", package = "climatrends")

# some random dates provided as integers and coerced to Dates internally
set.seed(2718279)
dates <- as.integer(runif(5, 17660, 17675))

# get precipitation indices for 30 days after day.one
# return a data.frame
rain1 <- rainfall(lonlatsf,
                  day.one = dates,
                  span = 30,
                  as.sf = FALSE)
rain1

# get precipitation indices from "2010-12-01" to "2011-01-31"
rain2 <- rainfall(lonlatsf,
                  day.one = "2010-12-01",
                  last.day = "2011-01-31",
                  as.sf = FALSE)
rain2

# indices from "2010-12-01" to "2011-01-31" with intervals of 7 days
rain3 <- rainfall(lonlatsf,
                  day.one = "2010-12-01",
                  last.day = "2011-01-31",
                  timeseries = TRUE,
                  intervals = 7,
                  as.sf = FALSE)
rain3

}       
         
}
\references{
Aguilar E., et al. (2005). Journal of Geophysical Research, 
110(D23), D23107. \cr\url{https://doi.org/10.1029/2005JD006119}
}
\concept{climatology functions}
\concept{precipitation functions}
