% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/climatrends.R
\docType{package}
\name{climatrends}
\alias{climatrends}
\alias{climatrends-package}
\title{Precipitation and Temperature Indices for Climate Variability Analysis}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

Toolkit for climate variability indices for ecological modelling. 
    The indices produced here can be used as explanatory variables for crop 
    modelling, trends in climate change and to assess the interactions
    of plants and animals with the environment.
}
\seealso{
\strong{Useful links:}
\itemize{
\item{Development repository: 
 \url{https://github.com/agrobioinfoservices/climatrends}}
\item{Static documentation: 
 \url{https://agrobioinfoservices.github.io/climatrends/}}
\item{Report bugs: 
 \url{https://github.com/agrobioinfoservices/climatrends/issues}}
}
}
\author{
Kauê de Sousa and Jacob van Etten and Svein Ø. Solberg
}
