% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_timeseries.R
\name{get_timeseries}
\alias{get_timeseries}
\alias{get_timeseries.default}
\alias{get_timeseries.sf}
\alias{get_timeseries.matrix}
\title{Time series climate data}
\usage{
get_timeseries(object, day.one, span, ...)

\method{get_timeseries}{default}(object, day.one, span, source = "nasapower", ...)

\method{get_timeseries}{sf}(object, day.one, span, source = "nasapower", ...)

\method{get_timeseries}{matrix}(object, day.one, span, ...)
}
\arguments{
\item{object}{a \code{data.frame} (or object that can be coerced to data.frame) 
with geographical coordinates (lonlat), or an object of class \code{sf} with 
geometry 'POINT' or 'POLYGON', or a named \code{matrix} with climate data.
See details.}

\item{day.one}{a vector of class \code{Date} for the starting date to capture 
the climate data}

\item{span}{an integer or a vector with integers for the duration of the 
time series to be captured}

\item{...}{additional arguments passed to methods. See details.}

\item{source}{character, for the source of climate data. See details.}
}
\value{
An object with time series climate data for the chosen period
}
\description{
General functions and methods to concatenate climate data across a time series.
}
\details{
The \code{default} method and the \code{sf} method assumes that the climate 
 data will be fetched from an remote (cloud) \var{source}.

The \code{matrix} method assumes that the climate data was previously handled 
 and will be inputted in the format of a named matrix. 
 See help("modis", "climatrends") for examples.

Available remote \var{source}s are: "nasapower"

Additional arguments:

\code{pars}: character vector of solar, meteorological or climatology parameters 
to download. See help("parameters", "nasapower") when \var{source} = "nasapower".

\code{days.before}: an integer for the number of days before \var{day.one} to be 
 included in the timespan.
}
\examples{
# Using local sources
data("modis", package = "climatrends")

day <- as.Date("2013-10-28", format = "\%Y-\%m-\%d")

get_timeseries(chirp, day, span = c(10,11))


########################################################
\donttest{
# Fetch data from NASA POWER using 'sf' method

data("lonlatsf", package = "climatrends")

do <- as.Date(17667, origin = "1970-01-01")

g <- get_timeseries(lonlatsf, 
                    day.one = do, 
                    span = 10,
                    pars = c("PRECTOT", "T2M", "T10M"))

g
}
}
\concept{GET functions}
