% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/late_frost.R
\name{late_frost}
\alias{late_frost}
\alias{late_frost.default}
\alias{late_frost.data.frame}
\alias{late_frost.array}
\alias{late_frost.sf}
\alias{late_frost.clima_ls}
\title{Late spring frost}
\usage{
late_frost(..., base = 4, tfrost = -2)

\method{late_frost}{default}(tmax, tmin, dates = NULL, ..., base = 4, tfrost = -2)

\method{late_frost}{data.frame}(object, day.one, ..., base = 4, tfrost = -2)

\method{late_frost}{array}(object, day.one, ..., base = 4, tfrost = -2)

\method{late_frost}{sf}(object, day.one, ..., base = 4, tfrost = -2)

\method{late_frost}{clima_ls}(object, ..., base = 4, tfrost = -2)
}
\arguments{
\item{...}{additional arguments passed to methods. See details}

\item{base}{an integer for the minimum temperature for growth}

\item{tfrost}{an integer for the freezing threshold}

\item{tmax}{a numeric vector with the maximum temperature}

\item{tmin}{a numeric vector with the minimum temperature}

\item{dates}{optional, a character (or Date or numeric) vector for the dates in 
\var{tmax} and \var{tmin}}

\item{object}{a data.frame with geographical coordinates (lonlat), 
or an object of class \code{sf} with geometry 'POINT' or 'POLYGON', 
or an \code{array} with two dimensions containing the temperature data 
or a \code{clima_ls} with maximum and minimum temperature, 
in that order. See details}

\item{day.one}{a vector of class \code{Date} or any other object that can be 
coerced to \code{Date} (e.g. integer, character YYYY-MM-DD) for the starting 
day to capture the climate data}
}
\value{
a data.frame with the late frost events
\item{id}{the id generated using the indices for the rows in \var{object}}
\item{date}{the first day of the event}
\item{gdd}{the growing degree-days accumulated during the event}
\item{event}{a factor for the observed event, frost, latent (where there is no frost event, 
 but also there is no GDD), and warming (where GDD is accumulated)}
\item{duration}{the number of days the event spanned}
}
\description{
Compute late spring frost, which is a freezing event occurring after
 a substantial accumulation of warmth
}
\details{
Additional arguments:

\code{equation}: character to specify the equation to be used, \code{"variant_a"} 
 is set by default. See GDD() 

\code{last.day}: an object (optional to \var{span}) of class \code{Date} or
 any other object that can be coerced to \code{Date} (e.g. integer, character 
 YYYY-MM-DD) for the last day of the time series
 
\code{span}: an integer (optional to \var{last.day}) or a vector with 
 integers (optional if \var{last.day} is given) for the length of 
 the time series to be captured
}
\examples{

# default method
data("innlandet", package = "climatrends")

tmax <- innlandet[, "tmax"]
tmin <- innlandet[, "tmin"]
date <- innlandet[, "dates"]

late_frost(tmax, tmin, dates = date)

# demo of the array method but no frost event is returned 
# because the data comes from the tropics
data("temp_dat", package = "climatrends")

late_frost(temp_dat, day.one = "2013-10-27")


\donttest{
# Some random points in Norway
# get data from NASA Power
library("nasapower")
data("lonlatsf", package = "climatrends")
  
late_frost(lonlatsf, day.one = "2019-01-01", last.day = "2019-07-01")
  
}

}
\references{
Trnka et al. (2014). Nature Climate Change 4(7):637–43.
\cr\url{https://doi.org/10.1038/nclimate2242}

Zohner et al. (2020). PNAS.
\cr\url{https://doi.org/10.1073/pnas.1920816117}
}
\seealso{
Other GDD functions: 
\code{\link{GDD}()}
}
\concept{GDD functions}
