% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rainfall.R
\name{rainfall}
\alias{rainfall}
\alias{rainfall.default}
\alias{rainfall.matrix}
\alias{rainfall.sf}
\title{Rainfall indices}
\usage{
rainfall(object, day.one, span, ...)

\method{rainfall}{default}(object, day.one, span, timeseries = FALSE, intervals = 5, ...)

\method{rainfall}{matrix}(object, day.one, span, timeseries = FALSE, intervals = 5, ...)

\method{rainfall}{sf}(
  object,
  day.one,
  span,
  timeseries = FALSE,
  intervals = 5,
  as.sf = TRUE,
  ...
)
}
\arguments{
\item{object}{a \code{data.frame} (or object that can be coerced to data.frame) 
with geographical coordinates (lonlat), or an object of class \code{sf} with 
geometry 'POINT' or 'POLYGON', or a named \code{matrix} with climate data.
See details.}

\item{day.one}{a vector of class \code{Date} for the starting date to capture 
the climate data}

\item{span}{an integer or a vector with integers for the duration of the 
time series to be captured}

\item{...}{additional arguments passed to methods. See details.}

\item{timeseries}{logical, \code{FALSE} for a single point time series
observation or \code{TRUE} for a time series based on \var{intervals}}

\item{intervals}{integer (no lower than 5), for the days intervals when
\var{timeseries} = \code{TRUE}}

\item{as.sf}{logical, to return an object of class 'sf'}
}
\value{
A dataframe with rainfall indices:
\item{MLDS}{maximum length of consecutive dry day, rain < 1 mm (days)}
\item{MLWS}{maximum length of consecutive wet days, rain >= 1 mm (days)}
\item{R10mm}{number of heavy precipitation days 10 >= rain < 20 mm (days)}
\item{R20mm}{number of very heavy precipitation days rain >= 20 (days)}
\item{Rx1day}{maximum 1-day precipitation (mm)}
\item{Rx5day}{maximum 5-day precipitation (mm)}
\item{R95p}{total precipitation when rain > 95th percentile (mm)}
\item{R99p}{total precipitation when rain > 99th percentile (mm)}
\item{Rtotal}{total precipitation (mm) in wet days, rain >= 1 (mm)}
\item{SDII}{simple daily intensity index, total precipitation divided by the
 number of wet days (mm/days)}
}
\description{
Rainfall indices over a timespan
}
\details{
The \code{matrix} method assumes that \var{object} contains climate data provided 
 from a local source; see help("chirp", package = "climatrends") for an example on input 
 structure.

The default method and the sf method assumes that the climate data will be fetched 
 from an remote (cloud) \var{source}.

When \var{timeseries} = \code{TRUE}, an id is created, 
 which is the index for the rownames of the inputted \var{object}.

Additional arguments:

\code{source}: character for the source of remote data. Current remote \var{source} 
 is: 'nasapower'

\code{pars}: character vector for the precipitation data to be fetched. If 
 \code{source} is 'nasapower', the default precipitation parameter is "PRECTOT".

\code{days.before}: optional, an integer for the number of days before 
 \var{day.one} to be included in the timespan.
 
When \var{timeseries} = \code{TRUE}, an id is created, 
 which is the index for the rownames of the inputted \var{object}.
}
\examples{
# Using local sources
data("chirp", package = "climatrends")

day <- as.Date("2013-10-28", format = "\%Y-\%m-\%d")

rainfall(chirp,
         day.one = day,
         span = 11)

###################################
\donttest{
########################################

# Using remote sources of climate data

data("lonlatsf", package = "climatrends")

set.seed(123)
dates <- as.integer(runif(5, 17660, 17675))
dates <- as.Date(dates, origin = "1970-01-01")

# get rainfall indices for 30 days after day.one
# return a data.frame
rain1 <- rainfall(lonlatsf,
                  day.one = dates,
                  span = 30, 
                  as.sf = FALSE)
rain1

# return a sf object
rain2 <- rainfall(lonlatsf,
                  day.one = dates,
                  span = 30)

rain2

# indices with intervals of 7 days and return a sf object 
rain3 <- rainfall(lonlatsf,
                  day.one = dates,
                  span = 30,
                  as.sf = FALSE,
                  timeseries = TRUE,
                  intervals = 7)
rain3

}       
         
}
\references{
Aguilar E., et al. (2005). Journal of Geophysical Research, 
110(D23), D23107. \cr\url{https://doi.org/10.1029/2005JD006119}
}
\concept{climatology functions}
\concept{precipitation functions}
