% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temperature.R
\name{temperature}
\alias{temperature}
\alias{temperature.default}
\alias{temperature.array}
\alias{temperature.sf}
\title{Temperature indices}
\usage{
temperature(object, day.one, span, ...)

\method{temperature}{default}(object, day.one, span, timeseries = FALSE, intervals = 5, ...)

\method{temperature}{array}(object, day.one, span, timeseries = FALSE, intervals = 5, ...)

\method{temperature}{sf}(
  object,
  day.one,
  span,
  timeseries = FALSE,
  intervals = 5,
  as.sf = TRUE,
  ...
)
}
\arguments{
\item{object}{a data.frame (or object that can be coerced to data.frame) 
with geographical coordinates (lonlat), or an object of class \code{sf}
with geometry 'POINT' or 'POLYGON', or an \code{array} with two dimensions 
containing the temperature data, see details.}

\item{day.one}{a vector of class \code{Date} for the starting date to capture 
the climate data}

\item{span}{an integer or a vector with integers for the duration of the 
time series to be captured}

\item{...}{additional arguments passed to methods. See details.}

\item{timeseries}{logical, \code{FALSE} for a single point time series
observation or \code{TRUE} for a time series based on \var{intervals}}

\item{intervals}{integer (no lower than 5), for the days intervals when
\var{timeseries} = \code{TRUE}}

\item{as.sf}{logical, to return an object of class 'sf'}
}
\value{
A dataframe with temperature indices:
\item{maxDT}{maximun day temperature (degree Celsius)}
\item{minDT}{minimum day temperature (degree Celsius)}
\item{maxNT}{maximun night temperature (degree Celsius)}
\item{minNT}{minimum night temperature (degree Celsius) }
\item{DTR}{diurnal temperature range (mean difference between DT 
and NT (degree Celsius)) }
\item{SU}{summer days, number of days with maximum temperature > 
30 (degree Celsius)}
\item{TR}{tropical nights, number of nights with maximum 
temperature > 25 (degree Celsius)}
\item{CFD}{consecutive frosty days, number of days with temperature 
bellow 0 degree Celsius}
\item{maxWSD}{maximum warm spell duration, consecutive days with 
temperature > 90th percentile}
\item{maxCSD}{maximum cold spell duration, consecutive nights with 
temperature < 10th percentile}
}
\description{
Methods to compute temperature indices over a time period
}
\details{
The \code{array} method assumes that \var{object} contains climate data provided 
 from a local source; this requires a array with two dimensions, 1st dimension 
 contains the day temperature and 2nd dimension the night temperature, 
 see help("modis", package = "climatrends") for an example on input structure.

The default method and the sf method assumes that the climate data will be fetched 
 from an remote (cloud) \var{source}.

When \var{timeseries} = \code{TRUE}, an id is created, 
 which is the index for the rownames of the inputted \var{object}.

Additional arguments:

\code{source}: character for the source of remote data. Current remote \var{source} 
 is: 'nasapower'

\code{pars}: character vector for the temperature data to be fetched. If 
 \code{source} is 'nasapower', the temperature can be adjusted to 2 m, the default,
 c("T2M_MAX", "T2M_MIN") or 10 m c("T10M_MAX", "T10M_MIN") 

\code{days.before}: optional, an integer for the number of days before 
 \var{day.one} to be included in the timespan.
}
\examples{
# Using local sources
data("modis", package = "climatrends")

day <- as.Date("2013-10-28", format = "\%Y-\%m-\%d")

temperature(modis, 
            day.one = day,
            span = 12)

\donttest{
#####################################################
 
# Using remote sources of climate data

data("lonlatsf", package = "climatrends")

set.seed(123)
dates <- as.integer(runif(5, 17660, 17675))
dates <- as.Date(dates, origin = "1970-01-01")

# get temperature indices for 30 days after day.one
# return a data.frame
temp1 <- temperature(lonlatsf,
                    day.one = dates,
                    span = 30,
                    as.sf = FALSE)
temp1

# return a sf object
temp2 <- temperature(lonlatsf,
                    day.one = dates,
                    span = 30)
temp2

# indices with intervals of 7 days and return a sf object 
temp3 <- temperature(lonlatsf,
                    day.one = dates,
                    span = 30,
                    timeseries = TRUE,
                    intervals = 7)
temp3

}
}
\references{
Aguilar E., et al. (2005). Journal of Geophysical Research, 
110(D23), D23107. \cr\url{https://doi.org/10.1029/2005JD006119}
}
\seealso{
Other temperature functions: 
\code{\link{ETo}()},
\code{\link{GDD}()}
}
\concept{temperature functions}
